优化wine和chrome支持离线安装

+++ src/ubuntu/install/chrome/install_chrome.sh
@@ -14,7 +14,9 @@ if [[ "${DISTRO}" == @(centos|oracle8|rockylinux9|rockylinux8|oracle9|almalinux9
   if [ ! -z "${CHROME_VERSION}" ]; then
     wget https://dl.google.com/linux/chrome/rpm/stable/x86_64/google-chrome-stable-${CHROME_VERSION}.x86_64.rpm -O chrome.rpm
   else
-    wget https://dl.google.com/linux/direct/google-chrome-stable_current_x86_64.rpm -O chrome.rpm
+    if [ ! -f "/opt/chrome.rpm" ]; then
+      wget https://dl.google.com/linux/direct/google-chrome-stable_current_x86_64.rpm -O chrome.rpm
+    fi
   fi
   if [[ "${DISTRO}" == @(oracle8|rockylinux9|rockylinux8|oracle9|almalinux9|almalinux8) ]]; then
     dnf localinstall -y chrome.rpm
@@ -42,7 +44,9 @@ else
   if [ ! -z "${CHROME_VERSION}" ]; then
     wget https://dl.google.com/linux/chrome/deb/pool/main/g/google-chrome-stable/google-chrome-stable_${CHROME_VERSION}_amd64.deb -O chrome.deb
   else
-    wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb -O chrome.deb
+    if [ ! -f "/opt/chrome.deb" ]; then
+      wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb -O chrome.deb
+    fi
   fi
   apt-get install -y ./chrome.deb
   rm chrome.deb
